refactor!: move objects into multiple modules
Also, add From implementations for some enums.

For categories in the Bot API documentation I did not provide top level
re-exports. They are separated on the website so they can also be
imported from their own module by users.

files except trait_*, api_params and objects should be sorted like they
are on the Bot API Documentation Website.

BREAKING CHANGE: imports change

BREAKING CHANGE: Media renamed to MediaGroupInputMedia
